add AddFlagsComponentObject
