+++ config_sample.yml
@@ -4,7 +4,7 @@
 
 browser: CHROME #acceptable values: FF, CHROME, GHOST, ANDROID
 wikiName: mediawiki119 #default testing wiki
-env: prod #acceptable values: prod, preview, sandbox-number dev-name
+env: preview #acceptable values: prod, preview, sandbox-number dev-name
 credentialsPath: c:/selenium-config/config.xml #Path to your config.xml file
 captchaPath: c:/selenium-config/captcha.txt #Path to your captcha.txt file
 qs: null
