+++ src/test/java/com/wikia/webdriver/common/contentpatterns/URLsContent.java
@@ -67,6 +67,7 @@ public class URLsContent {
   public static final String SPECIAL_EDIT_HUB = "wiki/Special:EditHub";
   public static final String SPECIAL_MAPS = "wiki/Special:Maps";
   public static final String SPECIAL_VIDEO_PAGE_ADMIN = "wiki/Special:VideoPageAdmin/edit";
+  public static final String SPECIAL_WHAT_LINKS_HERE = "Special:WhatLinksHere";
 
   //Urls for mobile
   public static final String MOBILETEST_MAINPAGE = "wiki/Mobileregressiontesting_Wiki";
@@ -129,7 +130,7 @@ public class URLsContent {
   public static final String ARTICLENAME_002 = "TestVid002";
   public static final String ARTICLENAME_003 = "TestVid003";
   public static final String ARTICLENAME_004 = "TestVid004";
-  public static final String FlagTest = "FlagTest";
+  public static final String FLAGTEST = "FlagTest";
 
   public static final String SPECIAL_SEARCH = "wiki/Special:Search";
